Promote AccessibilitySnapshotModel to a standalone package and test it on macOS
Move the portable model into its own dependency-free SwiftPM package consumed by
the host package via a local path dependency. Its graph contains no UIKit, so its
Codable/wire-format tests run under a plain `swift test` on macOS (and Linux)
instead of only through the iOS simulator. The pure model tests move out of the
iOS unit test target; the UIKit-bound ones (UIAccessibilityTraits, UIBezierPath)
stay.
